+++ bundles/org.eclipse.wst.xml.xpath2.processor/src/org/eclipse/wst/xml/xpath2/processor/internal/function/FnHoursFromDuration.java
@@ -18,6 +18,7 @@ import org.eclipse.wst.xml.xpath2.processor.ResultSequenceFactory;
 import org.eclipse.wst.xml.xpath2.processor.internal.*;
 import org.eclipse.wst.xml.xpath2.processor.internal.types.*;
 
+import java.math.BigInteger;
 import java.util.*;
 
 /**
@@ -77,7 +78,7 @@ public class FnHoursFromDuration extends Function {
 		if (dtd.negative())
 			res *= -1;
 
-		rs.add(new XSInteger(res));
+		rs.add(new XSInteger(BigInteger.valueOf(res)));
 
 		return rs;
 	}
