+++ bundles/org.eclipse.wst.xml.xpath2.processor/src/org/eclipse/wst/xml/xpath2/processor/internal/function/OpTo.java
@@ -6,7 +6,8 @@
  * http://www.eclipse.org/legal/epl-v10.html
  *
  * Contributors:
- *     Andrea Bittau - initial API and implementation from the PsychoPath XPath 2.0 
+ *     Andrea Bittau - initial API and implementation from the PsychoPath XPath 2.0
+ *     Mukul Gandhi - bug 274805 - improvements to xs:integer data type 
  *******************************************************************************/
 
 package org.eclipse.wst.xml.xpath2.processor.internal.function;
@@ -17,6 +18,7 @@ import org.eclipse.wst.xml.xpath2.processor.ResultSequenceFactory;
 import org.eclipse.wst.xml.xpath2.processor.internal.*;
 import org.eclipse.wst.xml.xpath2.processor.internal.types.*;
 
+import java.math.BigInteger;
 import java.util.*;
 
 /**
@@ -69,15 +71,15 @@ public class OpTo extends Function {
 		// get arguments
 		Iterator iter = cargs.iterator();
 		ResultSequence r = (ResultSequence) iter.next();
-		int one = ((XSInteger) r.first()).int_value();
+		int one = ((XSInteger) r.first()).int_value().intValue();
 		r = (ResultSequence) iter.next();
-		int two = ((XSInteger) r.first()).int_value();
+		int two = ((XSInteger) r.first()).int_value().intValue();
 
 		if (one > two)
 			return rs;
 
 		// inclusive first and last
-		rs.add(new XSInteger(one));
+		rs.add(new XSInteger(BigInteger.valueOf(one)));
 
 		if (one == two) {
 			return rs;
